LPD-87027 Move the Site Template sync flow to LayoutSetPrototypeHelperImpl
Pull the layout set prototype sync orchestration out of SitesImpl into
LayoutSetPrototypeHelperImpl, where it belongs conceptually. Sites
loses mergeLayoutSetPrototypeLayouts(Group, LayoutSet) and
isLayoutSetMergeable(Group, LayoutSet). The "set template without
enabling the link" one-shot import stays in site-impl alongside its
only caller.

# breaking

## What portal-kernel/src/com/liferay/sites/kernel/util/Sites.java

The public methods isLayoutSetMergeable(Group, LayoutSet) and mergeLayoutSetPrototypeLayouts(Group, LayoutSet) were removed from Sites.

## Why

The Site Template sync orchestration moved into LayoutSetPrototypeHelperImpl, where it belongs conceptually, and Sites is no longer a gateway for it.

----
