+++ modules/apps/layout/layout-set-prototype-impl/build.gradle
@@ -8,6 +8,7 @@ dependencies {
 	compileOnly project(":apps:export-import:export-import-api")
 	compileOnly project(":apps:layout:layout-page-template-api")
 	compileOnly project(":apps:layout:layout-set-prototype-api")
+	compileOnly project(":apps:portal-background-task:portal-background-task-api")
 	compileOnly project(":apps:xstream:xstream-configurator-api")
 	compileOnly project(":core:petra:petra-function")
 	compileOnly project(":core:petra:petra-sql-dsl-api")
