+++ modules/apps/layout/layout-set-prototype-impl/src/main/java/com/liferay/layout/set/prototype/internal/helper/LayoutSetPrototypeHelperImpl.java
@@ -5,13 +5,27 @@
 
 package com.liferay.layout.set.prototype.internal.helper;
 
+import com.liferay.exportimport.kernel.background.task.BackgroundTaskExecutorNames;
+import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationSettingsMapFactoryUtil;
+import com.liferay.exportimport.kernel.configuration.constants.ExportImportConfigurationConstants;
+import com.liferay.exportimport.kernel.lar.ExportImportHelper;
+import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
+import com.liferay.exportimport.kernel.lar.PortletDataHandlerKeys;
+import com.liferay.exportimport.kernel.lar.UserIdStrategy;
+import com.liferay.exportimport.kernel.model.ExportImportConfiguration;
+import com.liferay.exportimport.kernel.service.ExportImportConfigurationLocalService;
+import com.liferay.exportimport.kernel.service.ExportImportLocalService;
 import com.liferay.exportimport.kernel.staging.MergeLayoutPrototypesThreadLocal;
 import com.liferay.layout.set.prototype.helper.LayoutSetPrototypeHelper;
 import com.liferay.layout.set.prototype.sync.LayoutSetPrototypeSyncContextThreadLocal;
 import com.liferay.layout.set.prototype.sync.LayoutSetPrototypeSyncSessionManager;
 import com.liferay.petra.sql.dsl.DSLQueryFactoryUtil;
 import com.liferay.petra.sql.dsl.expression.Predicate;
+import com.liferay.portal.background.task.util.comparator.BackgroundTaskCreateDateComparator;
+import com.liferay.portal.kernel.backgroundtask.BackgroundTask;
+import com.liferay.portal.kernel.backgroundtask.BackgroundTaskManager;
 import com.liferay.portal.kernel.backgroundtask.constants.BackgroundTaskConstants;
+import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -25,6 +39,7 @@ import com.liferay.portal.kernel.model.LayoutSetPrototype;
 import com.liferay.portal.kernel.model.LayoutSetPrototypeTable;
 import com.liferay.portal.kernel.model.LayoutSetTable;
 import com.liferay.portal.kernel.model.LayoutTable;
+import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
@@ -35,17 +50,23 @@ import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.service.LayoutService;
 import com.liferay.portal.kernel.service.LayoutSetLocalService;
 import com.liferay.portal.kernel.service.LayoutSetPrototypeLocalService;
+import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.service.permission.GroupPermissionUtil;
 import com.liferay.portal.kernel.service.permission.LayoutPermission;
+import com.liferay.portal.kernel.util.LinkedHashMapBuilder;
 import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.uuid.PortalUUIDUtil;
 import com.liferay.sites.kernel.util.Sites;
 
+import java.io.Serializable;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -75,7 +96,7 @@ public class LayoutSetPrototypeHelperImpl implements LayoutSetPrototypeHelper {
 				_layoutSetLocalService.getLayoutSetsByLayoutSetPrototypeUuid(
 					layoutSetPrototype.getUuid())) {
 
-			if (_sites.isLayoutSetMergeable(layoutSet.getGroup(), layoutSet)) {
+			if (_isLayoutSetMergeable(layoutSet.getGroup(), layoutSet)) {
 				mergeableLayoutSets.add(layoutSet);
 			}
 		}
@@ -114,10 +135,38 @@ public class LayoutSetPrototypeHelperImpl implements LayoutSetPrototypeHelper {
 	}
 
 	@Override
-	public void executeLayoutSetSync(LayoutSet layoutSet) throws Exception {
+	public void executeLayoutSetSync(LayoutSet layoutSet)
+		throws PortalException {
+
 		MergeLayoutPrototypesThreadLocal.setSkipMerge(false);
 
-		_sites.mergeLayoutSetPrototypeLayouts(layoutSet.getGroup(), layoutSet);
+		if (MergeLayoutPrototypesThreadLocal.isSkipMerge()) {
+			_recordSyncResultIfTracked(BackgroundTaskConstants.STATUS_FAILED);
+
+			return;
+		}
+
+		MergeLayoutPrototypesThreadLocal.setSkipMerge(true);
+
+		Group group = layoutSet.getGroup();
+
+		layoutSet = _layoutSetLocalService.fetchLayoutSet(
+			layoutSet.getLayoutSetId());
+
+		if (!_isLayoutSetMergeable(group, layoutSet)) {
+			_recordSyncResultIfTracked(BackgroundTaskConstants.STATUS_FAILED);
+
+			return;
+		}
+
+		LayoutSetPrototype layoutSetPrototype =
+			_layoutSetPrototypeLocalService.
+				getLayoutSetPrototypeByUuidAndCompanyId(
+					layoutSet.getLayoutSetPrototypeUuid(),
+					layoutSet.getCompanyId());
+
+		_mergeLayoutSetPrototypeLayoutsInBackground(
+			layoutSet, layoutSetPrototype);
 	}
 
 	@Override
@@ -543,6 +592,69 @@ public class LayoutSetPrototypeHelperImpl implements LayoutSetPrototypeHelper {
 			));
 	}
 
+	private Map<String, String[]> _getLayoutSetPrototypesParameters() {
+		return LinkedHashMapBuilder.put(
+			PortletDataHandlerKeys.DATA_STRATEGY,
+			new String[] {PortletDataHandlerKeys.DATA_STRATEGY_MIRROR}
+		).put(
+			PortletDataHandlerKeys.DELETE_MISSING_LAYOUTS,
+			new String[] {Boolean.FALSE.toString()}
+		).put(
+			PortletDataHandlerKeys.DELETE_PORTLET_DATA,
+			new String[] {Boolean.FALSE.toString()}
+		).put(
+			PortletDataHandlerKeys.FAVICON,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.IGNORE_LAST_PUBLISH_DATE,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.LAYOUT_SET_SETTINGS,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.LAYOUT_SET_PROTOTYPE_LINK_ENABLED,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.LAYOUT_SET_PROTOTYPE_SETTINGS,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.LAYOUTS_IMPORT_MODE,
+			new String[] {
+				PortletDataHandlerKeys.
+					LAYOUTS_IMPORT_MODE_CREATED_FROM_PROTOTYPE
+			}
+		).put(
+			PortletDataHandlerKeys.LOGO, new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.PERMISSIONS,
+			new String[] {Boolean.FALSE.toString()}
+		).put(
+			PortletDataHandlerKeys.PORTLET_CONFIGURATION,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.PORTLET_CONFIGURATION_ALL,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.PORTLET_DATA,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.PORTLET_DATA_ALL,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.PORTLET_SETUP_ALL,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.THEME_REFERENCE,
+			new String[] {Boolean.TRUE.toString()}
+		).put(
+			PortletDataHandlerKeys.UPDATE_LAST_PUBLISH_DATE,
+			new String[] {Boolean.FALSE.toString()}
+		).put(
+			PortletDataHandlerKeys.USER_ID_STRATEGY,
+			new String[] {UserIdStrategy.CURRENT_USER_ID}
+		).build();
+	}
+
 	private boolean _hasDuplicatedFriendlyURLPrototypeLayout(
 			String layoutSetPrototypeLayoutERC, long groupId,
 			boolean privateLayout, String friendlyURL)
@@ -585,6 +697,167 @@ public class LayoutSetPrototypeHelperImpl implements LayoutSetPrototypeHelper {
 		return true;
 	}
 
+	private boolean _isLayoutSetMergeable(Group group, LayoutSet layoutSet) {
+		if (!layoutSet.isLayoutSetPrototypeLinkActive() ||
+			group.isLayoutPrototype() || group.isLayoutSetPrototype()) {
+
+			return false;
+		}
+
+		return true;
+	}
+
+	private boolean _isLayoutSetPrototypeMergeBackgroundTaskExists(
+		LayoutSetPrototype layoutSetPrototype, LayoutSet layoutSet) {
+
+		List<BackgroundTask> incompleteBackgroundTasks =
+			_backgroundTaskManager.getBackgroundTasks(
+				layoutSet.getGroupId(),
+				BackgroundTaskExecutorNames.
+					LAYOUT_SET_PROTOTYPE_MERGE_BACKGROUND_TASK_EXECUTOR,
+				false, QueryUtil.ALL_POS, QueryUtil.ALL_POS,
+				BackgroundTaskCreateDateComparator.getInstance(false));
+
+		for (BackgroundTask incompleteBackgroundTask :
+				incompleteBackgroundTasks) {
+
+			long exportImportConfigurationId = MapUtil.getLong(
+				incompleteBackgroundTask.getTaskContextMap(),
+				"exportImportConfigurationId");
+
+			ExportImportConfiguration exportImportConfiguration =
+				_exportImportConfigurationLocalService.
+					fetchExportImportConfiguration(exportImportConfigurationId);
+
+			if (exportImportConfiguration != null) {
+				Map<String, Serializable> settingsMap =
+					exportImportConfiguration.getSettingsMap();
+
+				Map<String, String[]> parameterMap =
+					(Map<String, String[]>)settingsMap.get("parameterMap");
+
+				long layoutSetId = MapUtil.getLong(parameterMap, "layoutSetId");
+
+				if (layoutSetId == layoutSet.getLayoutSetId()) {
+					if (incompleteBackgroundTask.getStatus() !=
+							BackgroundTaskConstants.STATUS_IN_PROGRESS) {
+
+						return true;
+					}
+
+					long lastMergeVersion = MapUtil.getLong(
+						parameterMap, "lastMergeVersion");
+
+					if (lastMergeVersion ==
+							layoutSetPrototype.getMvccVersion()) {
+
+						return true;
+					}
+				}
+			}
+		}
+
+		return false;
+	}
+
+	private void _mergeLayoutSetPrototypeLayoutsInBackground(
+			LayoutSet layoutSet, LayoutSetPrototype layoutSetPrototype)
+		throws PortalException {
+
+		if (ExportImportThreadLocal.isExportInProcess() ||
+			ExportImportThreadLocal.isImportInProcess() ||
+			ExportImportThreadLocal.isStagingInProcess()) {
+
+			_recordSyncResultIfTracked(BackgroundTaskConstants.STATUS_FAILED);
+
+			return;
+		}
+
+		if (_isLayoutSetPrototypeMergeBackgroundTaskExists(
+				layoutSetPrototype, layoutSet)) {
+
+			if (_log.isDebugEnabled()) {
+				_log.debug(
+					"Layout set prototype merge is in progress for layout " +
+						"set " + layoutSet.getLayoutSetId());
+			}
+
+			_recordSyncResultIfTracked(BackgroundTaskConstants.STATUS_FAILED);
+
+			return;
+		}
+
+		Map<String, String[]> parameterMap =
+			_getLayoutSetPrototypesParameters();
+
+		parameterMap.put(
+			PortletDataHandlerKeys.LAYOUT_SET_PRIVATE_LAYOUT,
+			new String[] {String.valueOf(layoutSet.isPrivateLayout())});
+		parameterMap.put(
+			"layoutSetId",
+			new String[] {String.valueOf(layoutSet.getLayoutSetId())});
+		parameterMap.put(
+			"layoutSetPrototypeId",
+			new String[] {
+				String.valueOf(layoutSetPrototype.getLayoutSetPrototypeId())
+			});
+
+		String syncSessionId =
+			LayoutSetPrototypeSyncContextThreadLocal.getSyncSessionId();
+
+		if (Validator.isNotNull(syncSessionId)) {
+			parameterMap.put("syncSessionId", new String[] {syncSessionId});
+		}
+
+		User user = _userLocalService.getDefaultUser(layoutSet.getCompanyId());
+
+		List<Layout> layoutSetPrototypeLayouts = _layoutLocalService.getLayouts(
+			layoutSetPrototype.getGroupId(), true);
+
+		Map<String, Serializable> exportLayoutSettingsMap =
+			ExportImportConfigurationSettingsMapFactoryUtil.
+				buildExportLayoutSettingsMap(
+					user, layoutSetPrototype.getGroupId(), true,
+					_exportImportHelper.getLayoutIds(layoutSetPrototypeLayouts),
+					parameterMap);
+
+		ExportImportConfiguration exportImportConfiguration = null;
+
+		try {
+			exportImportConfiguration =
+				_exportImportConfigurationLocalService.
+					addDraftExportImportConfiguration(
+						user.getUserId(),
+						ExportImportConfigurationConstants.TYPE_EXPORT_LAYOUT,
+						exportLayoutSettingsMap);
+		}
+		catch (PortalException portalException) {
+			_log.error(
+				"Unable to add draft export-import configuration",
+				portalException);
+
+			_recordSyncResultIfTracked(BackgroundTaskConstants.STATUS_FAILED);
+
+			return;
+		}
+
+		_exportImportLocalService.mergeLayoutSetPrototypeInBackground(
+			user.getUserId(), layoutSet.getGroupId(),
+			exportImportConfiguration);
+	}
+
+	private void _recordSyncResultIfTracked(int backgroundTaskStatus) {
+		String syncSessionId =
+			LayoutSetPrototypeSyncContextThreadLocal.getSyncSessionId();
+
+		if (Validator.isNull(syncSessionId)) {
+			return;
+		}
+
+		_layoutSetPrototypeSyncSessionManager.recordBackgroundTaskStatus(
+			backgroundTaskStatus, syncSessionId);
+	}
+
 	/**
 	 * Resets the modified timestamp on the layout so the linked page template is
 	 * merged into the layout when it is first accessed.
@@ -599,6 +872,19 @@ public class LayoutSetPrototypeHelperImpl implements LayoutSetPrototypeHelper {
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		LayoutSetPrototypeHelperImpl.class);
+
+	@Reference
+	private BackgroundTaskManager _backgroundTaskManager;
+
+	@Reference
+	private ExportImportConfigurationLocalService
+		_exportImportConfigurationLocalService;
+
+	@Reference
+	private ExportImportHelper _exportImportHelper;
+
+	@Reference
+	private ExportImportLocalService _exportImportLocalService;
 
 	@Reference
 	private GroupLocalService _groupLocalService;
@@ -626,6 +912,6 @@ public class LayoutSetPrototypeHelperImpl implements LayoutSetPrototypeHelper {
 		_layoutSetPrototypeSyncSessionManager;
 
 	@Reference
-	private Sites _sites;
+	private UserLocalService _userLocalService;
 
 }
